backend/requirements.txt
fastapi[all]
celery[redis]
redis
g4f
requests
pydantic
websockets
pillow
